+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.3.0]
+
+### Added
+
+- Better theme name resolution. e.g. "Nord Dark", "Nord-Dark", "nord-dark", "nord dark", "nordDark" all resolve to the same theme.
+- TcaTheme::default() is the same as TcaTheme::new(None)
+
 ## [0.2.0]
 
 ### Added